Shortlinks: Don't Show Jetpack Sidebar if Unavailable (#30663)
